+++ contracts/contracts/interfaces/ILongShort.sol
@@ -3,12 +3,12 @@
 pragma solidity 0.8.2;
 
 contract ILongShort {
-    function redeemLong(uint256 marketIndex, uint256 tokensToRedeem)
+    function redeemLong(uint32 marketIndex, uint256 tokensToRedeem)
         external
         virtual
     {}
 
-    function redeemShort(uint256 marketIndex, uint256 tokensToRedeem)
+    function redeemShort(uint32 marketIndex, uint256 tokensToRedeem)
         external
         virtual
     {}
