+++ contracts/contracts/interfaces/IStaker.sol
@@ -4,7 +4,7 @@ pragma solidity 0.8.2;
 
 abstract contract IStaker {
     function addNewStakingFund(
-        uint256 marketIndex,
+        uint32 marketIndex,
         address longTokenAddress,
         address shortTokenAddress,
         uint256 kInitialMultiplier,
